refactor(restaurantDetails): separate logic for restaurant details
