+++ src/components/pages/Restaurant/RestaurantsDetails/RestaurantsDetails.tsx
@@ -1,6 +1,4 @@
-import { useEffect, useState } from "react";
-import { useParams } from "react-router-dom";
-import { IRestaurantsDetails, Menu } from "../../../../static/interfaces";
+import { Menu } from "../../../../static/interfaces";
 import { ShoppingCart } from "../../../Cart/Cart";
 import {
   FilterBtnWrapper,
@@ -11,54 +9,11 @@ import { RestaurantWrapper } from "../Restaurant.style";
 import { MealHolder } from "../../../Meal/MealHolder";
 import { ClearAllFilter } from "../../../Menu/MenuFIlter/ClearAllFilter";
 import { clearFilter } from "../../../../static/endpoints";
-import { usePopupContext } from "../../../../context/PopupContext";
-import { mealService } from "../../../../services/mealService";
+import { useRestaurantDetailsLogic } from "./RestaurantsDetails.logic";
 
 export const RestaurantsDetails = () => {
-  const {
-    isUpdateMealPopUpVisible,
-    isUpdateMenuPopUpVisible,
-    isAddMealPopUpVisible,
-    isDeleteMealPopUpVisible,
-    isDeleteMenuPopUpVisible,
-    isDeleteRestaurantPopUpVisible,
-  } = usePopupContext();
-  const { id } = useParams();
-  const [restaurantDetails, setRestaurantDetails] =
-    useState<IRestaurantsDetails | null>(null);
-  const [menus, setMenus] = useState<Menu[] | null>(null);
-  const [allMenus, setAllMenus] = useState<Menu[] | null>(null);
-
-  const filter = (type: string) => {
-    if (type === clearFilter.all) {
-      setMenus(allMenus);
-    } else {
-      setMenus(allMenus?.filter((menu) => menu.type === type) || []);
-    }
-  };
-
-  useEffect(() => {
-    const fetchRestaurantDetails = async () => {
-      try {
-        const data = await mealService.fetchMealsByRestaurantId(id);
-        setRestaurantDetails(data);
-        setMenus(data?.menus || []);
-        setAllMenus(data?.menus || []);
-      } catch (error) {
-        console.error("Error fetching restaurant details:", error);
-      }
-    };
-
-    fetchRestaurantDetails();
-  }, [
-    id,
-    isUpdateMealPopUpVisible,
-    isUpdateMenuPopUpVisible,
-    isAddMealPopUpVisible,
-    isDeleteMealPopUpVisible,
-    isDeleteMenuPopUpVisible,
-    isDeleteRestaurantPopUpVisible,
-  ]);
+  const { restaurantDetails, filter, menus, allMenus } =
+    useRestaurantDetailsLogic();
 
   if (!restaurantDetails) {
     return <div>Loading...</div>;
